==> 2. Cleaning_PCD_Data_After_for_Visualization.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

%cd '/content/drive/MyDrive/Project_2/Project2_ใจผมสลายฮะมุง/source_code'

# 0. Import Library

import pandas as pd

# 1. Import Data

path = 'cleaned_data/' # Change path here
filename = 'cleaned_pcd_data_after.csv'

df = pd.read_csv(path+filename)
df['date_time'] = pd.to_datetime(df['date_time'])

df

df.info()

# 2.1 Create dataframe for visualization

cleaned_df = pd.DataFrame()

province_array = df['province'].unique()
month_period = pd.period_range(start='2020-07', end='2022-02', freq='M').astype(str)

month_period

for province in province_array:
  cleaned_df_new = pd.DataFrame()
  print('Province: '+province)
  selected_df = df[df['province']==province][['date_time', 'lat', 'long', 'PM25']]
  lat_value = selected_df.loc[selected_df.index[0], 'lat']
  long_value = selected_df.loc[selected_df.index[0], 'long']
  cleaned_df_new['date'] = pd.period_range(start='2020-07', end='2022-01', freq='M')
  pm25_mean = []
  for i in range(len(month_period)-1):
    pm25_mean.append(selected_df[(selected_df['date_time']<month_period[i+1]) & (selected_df['date_time']>=month_period[i])]['PM25'].mean())
  cleaned_df_new['PM25'] = pm25_mean
  # Fill first row of each station with 0 if it has NaN value.
  cleaned_df_new.loc[0] = cleaned_df_new.loc[0].fillna(0)
  # Fill missing values with its previous value (avoid using future data)
  cleaned_df_new = cleaned_df_new.interpolate(method = 'pad', limit_direction = 'forward') # Change interpolate method here
  prov_list = [province]*(len(month_period)-1)
  cleaned_df_new['province'] = prov_list
  lat_list = [lat_value]*(len(month_period)-1)
  cleaned_df_new['lat'] = lat_list
  long_list = [long_value]*(len(month_period)-1)
  cleaned_df_new['long'] = long_list
  cleaned_df = pd.concat([cleaned_df, cleaned_df_new], ignore_index=True)

cleaned_df

cleaned_df = cleaned_df.sort_values(by=['date'])
cleaned_df = cleaned_df.reset_index(drop=True)

cleaned_df

save_path = 'cleaned_data/'
save_filename = 'cleaned_pcd_data_after_for_visualization.csv'

cleaned_df.to_csv(save_path+save_filename, index=False)

# 2.2 EDA Analysis

cleaned_df = pd.DataFrame()
cleaned_df['date'] = pd.period_range(start='2020-07', end='2022-01', freq='M')

province_array = df['province'].unique()
month_period = pd.period_range(start='2020-07', end='2022-02', freq='M').astype(str)

month_period

cleaned_df

for province in province_array:
  print('Province: '+province)
  selected_df = df[df['province']==province][['date_time', 'PM25']]
  pm25_mean = []
  for i in range(len(month_period)-1):
    pm25_mean.append(selected_df[(selected_df['date_time']<month_period[i+1]) & (selected_df['date_time']>=month_period[i])]['PM25'].mean())
  cleaned_df[province] = pm25_mean

selected_df = df[df['province']=='Nakhon Nayok'][['date_time', 'PM25']]

selected_df

month_period[i]

(selected_df['date_time']<month_period[i+1])

(selected_df['date_time']>=month_period[i])

selected_df[(selected_df['date_time']<month_period[i+1]) & (selected_df['date_time']>=month_period[i])]['PM25'].mean()

cleaned_df

# 3. Impute Missing value

# Fill first row of each station with 0 if it has NaN value.
cleaned_df.loc[0] = cleaned_df.loc[0].fillna(0)
# Fill missing values with its previous value (avoid using future data)
cleaned_df = cleaned_df.interpolate(method = 'pad', limit_direction = 'forward') # Change interpolate method here

cleaned_df.head()

# 4. Save cleaned dataframe to csv file

save_path = 'cleaned_data/'
save_filename = 'cleaned_pcd_data_after_for_eda.csv'

cleaned_df.to_csv(save_path+save_filename, index=False)
