==> 7. Mapping_PCD_Data_Before_Station.ipynb <==
# **Import and Install Necessary Packages**

import os
import pandas as pd

# **Mount Google Drive**

from google.colab import drive
drive.mount('/content/drive')

#Change Directory Here
%cd /content/drive/MyDrive/my university life/YEAR 3 - JUNIOR/Data Science and Data Engineering/Project_2

# Mapping all PCD Data (Before) Files with Station

BASE_DATA_PATH = './data/PCD Data/Data before 2020-9'
STATION_METADATA = []
UNIQUE_STATION_IDS = set()

station_count = 0

#traverse all files inside data/PCD Data/Data before 2020-9
for path, folders, files in os.walk(BASE_DATA_PATH):
    for filename in files:
        if not (filename.endswith('.xlsx') or filename.endswith('.xls')):
            continue

        if filename.endswith('.xlsx'):
            stationFileType = '.xlsx'
            stationName = filename[filename.find(")")+1:-5]
        elif filename.endswith('.xls'):
            stationFileType = '.xls'
            stationName = filename[filename.find(")")+1:-4]

        stationPath = os.path.join(path, filename)
        stationID = filename[filename.find("(")+1:filename.find(")")]

        if stationID in UNIQUE_STATION_IDS:
            i = 1
            newStationID = stationID + '_' + str(i)
            while (newStationID in UNIQUE_STATION_IDS):
                i += 1
                newStationID = stationID + '_' + str(i)
            stationID = newStationID
        
        STATION_METADATA.append({
            'stationID': stationID,
            'stationName': stationName,
            'filePath': stationPath,
            'fileType': stationFileType
        })
        UNIQUE_STATION_IDS.add(stationID)
        station_count += 1

print(f'There are total of {station_count} stations.')

df = pd.DataFrame(STATION_METADATA)
df.sort_values(by = 'stationID', inplace = True)
df.reset_index(drop = True, inplace = True)
df

#save csv file to processed_data folder
BASE_PROCESS_PATH = './processed_data'
if not os.path.exists(BASE_PROCESS_PATH):
    os.mkdir(BASE_PROCESS_PATH)

df.to_csv(f'{BASE_PROCESS_PATH}/PCD_Data_Before_Stations.csv', index = False)
