==> 8. Mapping_Thailand_Province_With_Region_and_Coordinate.ipynb <==
# **Import Necessary Packages**

import pandas as pd

from selenium import webdriver
from selenium.webdriver.chrome.service import Service as ChromeService
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys

import time

# **Load Thai/English Province csv File**

df_province = pd.read_csv('./outsource_data/mapper-provinces-thailand.csv').iloc[:, 1:]
df_province['ProvinceTH'] = df_province['ProvinceTH'].str.replace('จ.', '', regex = False)
df_province.sort_values(by = 'ProvinceTH', inplace = True)

df_province

# **Load Thai/Region Province csv File**

df_region = pd.read_csv('./outsource_data/25640531_sc002_.csv').loc[:, ['ProvinceNameThai', 'Region_VaccineRollout_MOPH']]
df_region.rename(columns = {'ProvinceNameThai': 'ProvinceTH', 'Region_VaccineRollout_MOPH': 'Region'}, inplace = True)
df_region['ProvinceTH'] = df_region['ProvinceTH'].str.replace('จังหวัด', '', regex = False)
df_region.sort_values(by = 'ProvinceTH', inplace = True)

region_mapping = {
    'กรุงเทพมหานครและปริมณฑล': 'Bangkok Metropolitan',
    'ภาคใต้': 'Southern',
    'ภาคตะวันตก': 'Western',
    'ภาคตะวันออกเฉียงเหนือ': 'Northeastern',
    'ภาคกลาง': 'Central',
    'ภาคตะวันออก': 'Eastern',
    'ภาคเหนือ': 'Northern'
}
df_region['Region'] = df_region['Region'].map(region_mapping)

df_region

df_region['Region'].value_counts()

# **Merge Dataframes**

df_merged = pd.merge(df_province, df_region, on = 'ProvinceTH')
df_merged

# **Use ProvinceTH to find latitude, longitude from Google Maps using Selenium**|

url = 'https://www.google.co.th/maps?hl=en&tab=rl'
service = ChromeService(executable_path=ChromeDriverManager().install())
driver = webdriver.Chrome(service=service)
driver.get(url)
df_locations = []

for province in df_merged['ProvinceTH']:
    search_input = driver.find_element(By.CSS_SELECTOR, 'input#searchboxinput')
    search_button = driver.find_element(By.CSS_SELECTOR, '#searchbox-searchbutton')

    search_input.send_keys(province) #enter province name
    search_button.click() #click search button
    time.sleep(4) #wait for downloading
    current_url = driver.current_url
    location_data = current_url.split('/')[6].split(',')
    lat = float(location_data[0][1:])
    long = float(location_data[1])
    df_locations.append([province, lat, long])
    print([province, lat, long])
    driver.get(url)
driver.quit()
df_locations = pd.DataFrame(df_locations, columns = ['ProvinceTH', 'lat', 'long'])

df_locations

df_final_merged = pd.merge(df_merged, df_locations, on = 'ProvinceTH', how = 'left')

df_final_merged.to_csv('./processed_data/mapping_thailand_province_with_region_and_location.csv', index = False)
df_final_merged
