==> 9. paired_t_test.ipynb <==
import pandas as pd
import numpy as np

from scipy import stats
from statsmodels.stats import weightstats as stests

from google.colab import drive
drive.mount('/content/drive')

%cd '/content/drive/MyDrive'

df_before = pd.read_csv('cleaned_data/cleaned_pcd_data_before_all_parts.csv')
df_before['date_time'] = pd.to_datetime(df_before['date_time'], format = '%Y-%m-%d %H:%M')

df_before_station = pd.read_csv('./processed_data/PCD_DICTIONARY.csv')
df_before_station = df_before_station.iloc[:, 1:]
df_before_station.rename(columns = {'stationId': 'stationID', 'name': 'nameTH'}, inplace = True)

mapping_column = ['lat', 'long', 'province']
for c in mapping_column:
    mapping_dict = dict(df_before_station.set_index('stationID')[c])
    df_before[c] = df_before['stationID'].map(mapping_dict)
df_before['year'] = df_before['date_time'].dt.year
df_before

df1 = df_before[df_before['year'] == 2017]
df1

pm25_1_avg = df1.groupby('province')['PM2_5 at 3 m (microg/m3)'].mean().reset_index(name='pm2.5_1')
pm25_1_avg.head()

df2 = df_before[df_before['year'] == 2018]
df2

pm25_2_avg = df2.groupby('province')['PM2_5 at 3 m (microg/m3)'].mean().reset_index(name='pm2.5_2')
pm25_2_avg.head()

df3 = df_before[df_before['year'] == 2019]
df3

pm25_3_avg = df3.groupby('province')['PM2_5 at 3 m (microg/m3)'].mean().reset_index(name='pm2.5_3')
pm25_3_avg.head()

pm_avg = pm25_1_avg.merge(pm25_2_avg, on = 'province', how = 'inner')
pm_avg = pm_avg.merge(pm25_3_avg, on = 'province', how = 'inner')

pm_avg.dropna(subset=['pm2.5_1','pm2.5_2'], inplace=True)
pm_avg

def ttest_alt(a, b, alternative='two-sided'):
    tt, tp = stats.ttest_rel(a, b)

    if alternative == 'greater':
        if tt > 0:
            tp = 1 - (1-tp)/2
        else:
            tp /= 2
    elif alternative == 'less':
        if tt <= 0:
            tp /= 2
        else:
            tp = 1 - (1-tp)/2

    return tt, tp

ttest,pval = ttest_alt(pm_avg['pm2.5_1'], pm_avg['pm2.5_2'], alternative='less') #between2017,2018
print(pval)

ttest,pval = ttest_alt(pm_avg['pm2.5_2'], pm_avg['pm2.5_3'], alternative='less') #between2018,2019
print(pval)

ttest,pval = ttest_alt(pm_avg['pm2.5_1'], pm_avg['pm2.5_3'], alternative='less') #between2017,2019
print(pval)
